Cree un hook para el scroll así puede utilizarse en todos los componentes del home
